+++ app/src/main/java/com/mapbox/maps/testapp/examples/snapshotter/MapSnapshotterActivity.kt
@@ -1,7 +1,10 @@
 package com.mapbox.maps.testapp.examples.snapshotter
 
+import android.graphics.Paint
+import android.graphics.RectF
 import android.os.Bundle
 import android.widget.ImageView
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.mapbox.geojson.Point
 import com.mapbox.maps.*
@@ -34,12 +37,24 @@ class MapSnapshotterActivity : AppCompatActivity(), SnapshotStyleListener {
           )
         ).build()
       )
-      start {
-        it?.let { mapSnapshot ->
-          val imageView = ImageView(this@MapSnapshotterActivity)
-          imageView.setImageBitmap(mapSnapshot.bitmap())
-          setContentView(imageView)
+      start(
+        overlayCallback = { overlay ->
+          overlay.canvas.drawOval(
+            RectF(0f, 0f, 100f, 100f),
+            Paint().apply { alpha = 128 }
+          )
         }
+      ) { bitmap, errorMessage ->
+        if (errorMessage != null) {
+          Toast.makeText(
+            this@MapSnapshotterActivity,
+            errorMessage,
+            Toast.LENGTH_SHORT
+          ).show()
+        }
+        val imageView = ImageView(this@MapSnapshotterActivity)
+        imageView.setImageBitmap(bitmap)
+        setContentView(imageView)
       }
     }
   }